Add files via upload
The project is developed for the R.K. Bangali Library of Pratap College, Amalner. In this project librarian can add the records of the books. In this project Librarian can add multiple records of the books. Basically the project is built on client side.
This project is built by using HTML, CSS, Bootstrap, JavaScript.